Moved around Paginator tests.
