+++ tests/integration/Paginator/Adapter/QueryBuilder/PaginateCest.php
@@ -13,12 +13,18 @@ declare(strict_types=1);
 namespace Phalcon\Test\Integration\Paginator\Adapter\QueryBuilder;
 
 use IntegrationTester;
+use Phalcon\Test\Fixtures\Traits\DiTrait;
 
-/**
- * Class PaginateCest
- */
 class PaginateCest
 {
+    use DiTrait;
+
+    public function _before(IntegrationTester $I)
+    {
+        $this->setNewFactoryDefault();
+        $this->setDiMysql();
+    }
+
     /**
      * Tests Phalcon\Paginator\Adapter\QueryBuilder :: paginate()
      *
@@ -28,6 +34,7 @@ class PaginateCest
     public function paginatorAdapterQuerybuilderPaginate(IntegrationTester $I)
     {
         $I->wantToTest('Paginator\Adapter\QueryBuilder - paginate()');
+
         $I->skipTest('Need implementation');
     }
 }
